Refactor login and registration processes
Updated AccountController to use LoginUserByAuthenticationQuery in LoginUser method. Moved MatLidStoreDatabaseContext registration from IdentityServicesRegistration to PersistenceServiceRegistration. Replaced LoginModel with AuthRequest in login-layout.component.ts. Changed Phone to PhoneNumber in UserDtoValidator and RegistrationRequest, adding [Phone] attribute for validation. Added LoginUserByAuthenticationQuery.cs and updated handler. Removed LoginUserByUserNameQuery.cs.
