+++ src/MSL.Application/DTO/User/UserDtoValidator.cs
@@ -97,8 +97,8 @@ public class RegisterUserModelValidator : AbstractValidator<RegistrationRequest>
         RuleFor(x => x.LastName)
             .NotEmpty().WithMessage("LastName cannot be empty.")
             .MaximumLength(50).WithMessage("LastName must be less than 50 characters.");
-        RuleFor(x => x.Phone)
-            .NotEmpty().WithMessage("Phone cannot be empty.");
+        RuleFor(x => x.PhoneNumber)
+            .NotEmpty().WithMessage("PhoneNumber cannot be empty.");
     }
 }
