+++ src/MSL.Application/Features/User/Queries/LoginUserByAuthentication/LoginUserByAuthenticationQuery.cs
@@ -1,0 +1,6 @@
+﻿using MediatR;
+using MLS.Application.Models.Identity;
+
+namespace MLS.Application.Features.User.Queries.LoginUserByAuthentication;
+
+public record LoginUserByAuthenticationQuery(AuthRequest model) : IRequest<AuthResponse>;
