+++ src/MSL.Application/Features/User/Queries/LoginUserByAuthentication/LoginUserByAuthenticationQueryHandler.cs
@@ -4,20 +4,20 @@ using MLS.Application.Contracts.Logging;
 using MLS.Application.Features.User.Queries.GetUserDetails;
 using MLS.Application.Models.Identity;
 
-namespace MLS.Application.Features.User.Queries.LoginUserByUserName;
+namespace MLS.Application.Features.User.Queries.LoginUserByAuthentication;
 
-public class LoginUserByUserNameQueryHandler : IRequestHandler<LoginUserByUserNameQuery, AuthResponse>
+public class LoginUserByAuthenticationQueryHandler : IRequestHandler<LoginUserByAuthenticationQuery, AuthResponse>
 {
     private readonly IAuthService _authService;
     private readonly IAppLogger<GetUserDetailsQueryHandler> _logger;
 
-    public LoginUserByUserNameQueryHandler(IAuthService authService, IAppLogger<GetUserDetailsQueryHandler> logger)
+    public LoginUserByAuthenticationQueryHandler(IAuthService authService, IAppLogger<GetUserDetailsQueryHandler> logger)
     {
         _authService = authService;
         _logger = logger;
     }
 
-    public async Task<AuthResponse> Handle(LoginUserByUserNameQuery request, CancellationToken cancellationToken)
+    public async Task<AuthResponse> Handle(LoginUserByAuthenticationQuery request, CancellationToken cancellationToken)
     {
         var response = await _authService.Login(request.model);
