--- src/MSL.Application/Features/User/Queries/LoginUserByUserName/LoginUserByUserNameQuery.cs
@@ -1,6 +1,0 @@
-﻿using MediatR;
-using MLS.Application.Models.Identity;
-
-namespace MLS.Application.Features.User.Queries.LoginUserByUserName;
-
-public record LoginUserByUserNameQuery(AuthRequest model) : IRequest<AuthResponse>;
